pyproject.toml
[project]
name = "shopping-list-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "pytest-asyncio>=1.1.0",
    "sqlmodel>=0.0.24",
]
